% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_greatlakes_fishtissue.R
\name{map_greatlakes_fishtissue}
\alias{map_greatlakes_fishtissue}
\title{Great Lakes Map}
\usage{
map_greatlakes_fishtissue(
  data = NULL,
  sv = NULL,
  xlong = "Longitude",
  ylat = "Latitude",
  amount = "amount",
  lab_title = NULL,
  lab_subtitle = NULL,
  lab_caption = NULL
)
}
\arguments{
\item{data}{Data frame of all values.  Default = NULL}

\item{sv}{Screening value. Default = NULL}

\item{xlong}{Data column with longitude. Default = Longitude}

\item{ylat}{Data column with latitude. Default = Latitude}

\item{amount}{Data column with analyte value. Default = amount}

\item{lab_title}{Title string. Default = NULL}

\item{lab_subtitle}{Caption string. Default = NULL}

\item{lab_caption}{Caption string. Default = NULL}
}
\value{
A ggplot object is returned.
}
\description{
Map of Great Lakes with values above and below a screening
value.
}
\details{
Map of Great Lakes with two data layers.  Layer 1 are sites greater
than or equal to the screening value (exceedances).  Layer 2 are sites less
than the screening value (non-exceedances).

# 20230719, moved code to package scripts.  Derived for Great Lakes studies.
}
\examples{

data <- data_gl2010_map
sv <- 40
xlong <- "Longitude"
ylat <- "Latitude"
amount <- "Amount"
lab_title <- "Great Lakes, 2010"
lab_subtitle <- paste("PFOS", paste0("Screeing Value = ", sv), sep = "\n")
lab_caption <- NULL

map_greatlakes_fishtissue(data = data
                          , sv = sv
                          , xlong = xlong
                          , ylat = ylat
                          , amount = amount
                          , lab_title = lab_title
                          , lab_subtitle = lab_subtitle
                          , lab_caption = lab_caption)

}
