% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg_status.R
\name{msg_status}
\alias{msg_status}
\title{Status Message}
\usage{
msg_status(fun.status, fun.item.num.current, fun.item.num.total, fun.item.name)
}
\arguments{
\item{fun.status}{Status of item}

\item{fun.item.num.current}{Item number}

\item{fun.item.num.total}{Total number of items}

\item{fun.item.name}{Name of item}
}
\value{
Returns a text to the console
}
\description{
Print Status Message to Console
}
\details{
Function for informing user of progress

Text is composed using `print` so will have to flush console if used inside
of a function or loop.
}
