% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cdf_fishtissue.R
\name{plot_cdf_fishtissue}
\alias{plot_cdf_fishtissue}
\title{Plot Fish Study CDF}
\usage{
plot_cdf_fishtissue(
  fun.DF,
  fun.Main,
  fun.xlim.x,
  fun.xlim.y,
  fun.xlab,
  fun.ylab,
  fun.axis.4.at,
  fun.axis.4.labels,
  fun.axis.4.mtext,
  fun.SV.plot,
  fun.SV,
  fun.SV.int,
  fun.SV.lab,
  fun.SV2.plot,
  fun.SV2,
  fun.SV2.int,
  fun.SV2.lab,
  fun.SV3.plot,
  fun.SV3,
  fun.SV3.int,
  fun.SV3.lab,
  fun.SV4.plot,
  fun.SV4,
  fun.SV4.int,
  fun.SV4.lab,
  fun.break.boo = FALSE,
  fun.break.max.at = NA,
  fun.break.max.text = NA,
  fun.break.pos = NA,
  fun.break.axis.at = NA,
  fun.break.axis.labels = NA,
  leg_SV_str = "Screening Value",
  leg_SV_units = NA,
  fun.axis.2.mtext = NA,
  fun.xlog = FALSE
)
}
\arguments{
\item{fun.DF}{x}

\item{fun.Main}{x}

\item{fun.xlim.x}{x}

\item{fun.xlim.y}{x}

\item{fun.xlab}{x}

\item{fun.ylab}{x}

\item{fun.axis.4.at}{x}

\item{fun.axis.4.labels}{x}

\item{fun.axis.4.mtext}{x}

\item{fun.SV.plot}{x}

\item{fun.SV}{x}

\item{fun.SV.int}{x}

\item{fun.SV.lab}{x}

\item{fun.SV2.plot}{x}

\item{fun.SV2}{x}

\item{fun.SV2.int}{x}

\item{fun.SV2.lab}{x}

\item{fun.SV3.plot}{x}

\item{fun.SV3}{x}

\item{fun.SV3.int}{x}

\item{fun.SV3.lab}{x}

\item{fun.SV4.plot}{x}

\item{fun.SV4}{x}

\item{fun.SV4.int}{x}

\item{fun.SV4.lab}{x}

\item{fun.break.boo}{x}

\item{fun.break.max.at}{x}

\item{fun.break.max.text}{x}

\item{fun.break.pos}{x}

\item{fun.break.axis.at}{x}

\item{fun.break.axis.labels}{x}

\item{leg_SV_str}{x}

\item{leg_SV_units}{x}

\item{fun.axis.2.mtext}{x}

\item{fun.xlog}{x}
}
\value{

}
\description{
CDF for fish study projects
}
\details{
Plots CDF with different options for fish tissue studies (NLA,
NRSA, and Great Lakes).

Plots are base R.

Saves plots to a user defined folder.

uses the Hmisc and plotrix packages

# CDF plotting function for Fish Tissue project
# 20161005, Erik.Leppo@tetratech.com
# move function to its own script
# 20200401, edits
## Screening Value to "Human Health Benchmark", leg_SV_str
## add ng/g to parenthetical, leg_SV_units, adds if present
# 20210826, Add legend label update
# 20220322, Legend, "Data" to "Fillet Concentration Data"
# 20220414, move to package scripts, was fun.plot.CDF.fish
}
\examples{

}
