% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_upset_fishtissue.R
\name{plot_upset_fishtissue}
\alias{plot_upset_fishtissue}
\title{Plot Fish Study Upset}
\usage{
plot_upset_fishtissue(
  data,
  title_str = NA,
  title_x = 0.2,
  title_y = 0.9,
  title_fontsize = 16,
  cap_str = NA,
  cap_x = 0.5,
  cap_y = 0.025,
  cap_fontsize = 8
)
}
\arguments{
\item{data}{Data as data frame (variables are 1 or 0 as columns)}

\item{title_str}{Title string. Default = NA}

\item{title_x}{Title x placement.  Default = 0.2}

\item{title_y}{Title y placment. Default = 0.9}

\item{title_fontsize}{Title font size.  Default = 16}

\item{cap_str}{Caption string. Default = NA}

\item{cap_x}{Caption x placement. Default = 0.5}

\item{cap_y}{Caption y placement. Default = 0.025}

\item{cap_fontsize}{Caption font size. Default = 8}
}
\value{
A plot object is returned.
}
\description{
Upset plot for fish study projects
}
\details{
Upset plot with different options for fish tissue studies (NLA,
NRSA, and Great Lakes).

Plots use the UpSetR package.

Saves plots to a user defined folder.

# 20230413, moved code to package scripts.  Derived for Great Lakes studies.
}
