% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_xmin_fishtissue.R
\name{set_xmin_fishtissue}
\alias{set_xmin_fishtissue}
\title{Set X Minimum}
\usage{
set_xmin_fishtissue(fun.value)
}
\arguments{
\item{fun.value}{}
}
\value{
numeric value
}
\description{
Fish Tissue CDF Set Xmin
}
\details{
# 1. function for use in determining X min
Fish Tissue CDF plots
}
