% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cdf_fishtissue.R
\name{plot_cdf_fishtissue}
\alias{plot_cdf_fishtissue}
\title{Plot Fish Study CDF}
\usage{
plot_cdf_fishtissue(
  fun.DF,
  fun.Main,
  fun.xlim.x,
  fun.xlim.y,
  fun.xlab,
  fun.ylab,
  fun.axis.4.at,
  fun.axis.4.labels,
  fun.axis.4.mtext,
  fun.SV.plot,
  fun.SV,
  fun.SV.int,
  fun.SV.lab,
  fun.SV2.plot,
  fun.SV2,
  fun.SV2.int,
  fun.SV2.lab,
  fun.SV3.plot,
  fun.SV3,
  fun.SV3.int,
  fun.SV3.lab,
  fun.SV4.plot,
  fun.SV4,
  fun.SV4.int,
  fun.SV4.lab,
  fun.SV5.plot,
  fun.SV5,
  fun.SV5.int,
  fun.SV5.lab,
  fun.break.boo = FALSE,
  fun.break.max.at = NA,
  fun.break.max.text = NA,
  fun.break.pos = NA,
  fun.break.axis.at = NA,
  fun.break.axis.labels = NA,
  leg_SV_str = "Screening Value",
  leg_SV_units = NA,
  leg_SV_verbose = TRUE,
  fun.axis.2.mtext = NA,
  fun.xlog = FALSE
)
}
\arguments{
\item{fun.DF}{Data data frame}

\item{fun.Main}{plot(main)}

\item{fun.xlim.x}{plot(xlim), value 1}

\item{fun.xlim.y}{plot(xlim), value 2}

\item{fun.xlab}{plot(xlab)}

\item{fun.ylab}{plot(ylab)}

\item{fun.axis.4.at}{axis 4 (right), at}

\item{fun.axis.4.labels}{axis 4 (right), labels}

\item{fun.axis.4.mtext}{axis 4 (right), margin text label}

\item{fun.SV.plot}{Screening Value 1, boolean value to plot SV}

\item{fun.SV}{Screening Value 1, numeric value}

\item{fun.SV.int}{Screening Value 1, y-intercept value}

\item{fun.SV.lab}{Screening Value 1, legend (source)}

\item{fun.SV2.plot}{Screening Value 2, boolean value to plot SV}

\item{fun.SV2}{Screening Value 2, numeric value}

\item{fun.SV2.int}{Screening Value 2, y-intercept value}

\item{fun.SV2.lab}{Screening Value 2, legend (source)}

\item{fun.SV3.plot}{Screening Value 3, boolean value to plot SV}

\item{fun.SV3}{Screening Value 3, numeric value}

\item{fun.SV3.int}{Screening Value 3, y-intercept value}

\item{fun.SV3.lab}{Screening Value 3, legend (source)}

\item{fun.SV4.plot}{Screening Value 4, boolean value to plot SV}

\item{fun.SV4}{Screening Value 4, numeric value}

\item{fun.SV4.int}{Screening Value 4, y-intercept value}

\item{fun.SV4.lab}{Screening Value 4, legend (source)}

\item{fun.SV5.plot}{Screening Value 4, boolean value to plot SV}

\item{fun.SV5}{Screening Value 4, numeric value}

\item{fun.SV5.int}{Screening Value 4, y-intercept value}

\item{fun.SV5.lab}{Screening Value 4, legend (source)}

\item{fun.break.boo}{Should x-axis include a break. Default = FALSE}

\item{fun.break.max.at}{. Default = NA}

\item{fun.break.pos}{location of break. Default = NA}

\item{fun.break.axis.at}{break. Default = NA}

\item{fun.break.axis.labels}{new x-axis labels with break}

\item{leg_SV_str}{Legend screening value, SV text.  Default = "Screening Value"}

\item{leg_SV_units}{Legend screening value, units.  Default = NA}

\item{leg_SV_verbose}{Boolean to include SV in the legend with or without
lab, units, and SV value.  Default = TRUE (FALSE will be leg_SV_str only)}

\item{fun.axis.2.mtext}{axis 2 (left) margin text label. Default = NA}

\item{fun.xlog}{Boolean value if x-axis should be log10 transformed.  Default = FALSE}
}
\value{
CDF plot customized for Fish Tissue project.
}
\description{
CDF for fish study projects
}
\details{
Plots CDF with different options for fish tissue studies (NLA,
NRSA, and Great Lakes).

Plots are base R.

Saves plots to a user defined folder.

uses the Hmisc and plotrix packages

BREAKING CHANGE
2023-10-04
Added 5th Screening Value
If didn't specify parameters and assumed an order will break code

# CDF plotting function for Fish Tissue project
# 20161005, Erik.Leppo@tetratech.com
# move function to its own script
# 20200401, edits
## Screening Value to "Human Health Benchmark", leg_SV_str
## add ng/g to parenthetical, leg_SV_units, adds if present
# 20210826, Add legend label update
# 20220322, Legend, "Data" to "Fillet Concentration Data"
# 20220414, move to package scripts, was fun.plot.CDF.fish
# 20231006
## increase SV thickness 1.5 to 2 and color green to darkgreen
## trigger for SV annotations in legend

# Derived from NRSA version (more advanced with Log scale than GL version)
}
\examples{
# non at this time

}
